requirements.txt
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
llama-index
jupyterlab
playwright
pymupdf
wikipedia-api
llama-index-postprocessor-flag-embedding-reranker
llama-index-postprocessor-rankgpt-rerank
llama-index-postprocessor-cohere-rerank
llama-index-retrievers-bm25
ragas
langchain-openai
simplejson
unstructured
unstructured[md]
faiss-cpu
ollama
llama-index-llms-ollama
mlflow